Refs #33. Fixed some issues with choppinp and reducing.
