+++ src/lib/reduce_VULCAN.py
@@ -1268,168 +1268,6 @@ class ReduceVulcanData(object):
 
         return sorted(list(target_set))
 
-    def chop_reduce(self, chop_dir=None):
-        """
-        Chop and reduce
-        :return:
-        """
-        # check whether it is good to go
-        assert isinstance(self._reductionSetup, ReductionSetup), 'ReductionSetup is not correct.'
-        # configure the ReductionSetup
-        self._reductionSetup.process_configurations()
-
-        # get the chopped data directory if not specified
-        if chop_dir is None:
-            chop_dir = self._reductionSetup.get_reduced_data_dir()
-
-        message = 'Output GSAS files include:\n'
-
-        # get splitters workspaces
-        split_ws_name, split_info_table = self._reductionSetup.get_splitters(throw_not_set=True)
-
-        num_target_ws = self.get_number_chopped_ws(split_ws_name)
-        if num_target_ws > 20:
-            self.chop_and_reduce_large_output(split_ws_name, split_info_table)
-
-        # call SNSPowderReduction for chopping and
-        # TODO/ISSUE/FIXME/NOW/33 ----- Important! Modify it to **args
-        mantidsimple.SNSPowderReduction(Filename=self._reductionSetup.get_event_file(),
-                                        PreserveEvents=True,
-                                        CalibrationFile=self._reductionSetup.get_focus_file(),
-                                        CharacterizationRunsFile=self._reductionSetup.get_characterization_file(),
-                                        Binning="-0.001",
-                                        SaveAS="",
-                                        OutputDirectory=self._reductionSetup.get_gsas_dir(),
-                                        NormalizeByCurrent=False,
-                                        FilterBadPulses=0,
-                                        CompressTOFTolerance=0.,
-                                        FrequencyLogNames="skf1.speed",
-                                        WaveLengthLogNames="skf12.lambda",
-                                        SplittersWorkspace=split_ws_name,
-                                        SplitInformationWorkspace=split_info_table)
-
-        # create GSAS file for split workspaces
-        # convert the chopped data to GSAS file in VULCAN's special bin
-        info_table = AnalysisDataService.retrieve(split_info_table)
-        num_split_ws = info_table.rowCount()
-
-        target_chop_index_list = self.get_target_split_ws_index(split_ws_name)
-
-        everything_is_right = True
-
-        # find out a naming issue
-        run_str = '%d' % self._reductionSetup.get_run_number()
-        if chop_dir.count(run_str) == 0:
-            use_special_name = True
-        else:
-            use_special_name = False
-
-        chopped_ws_name_list = list()
-
-        print '[DB...BAT] Output workspace number: ', num_split_ws
-
-        for i_ws in range(num_split_ws):
-            # get the split workspace's name
-            ws_index = int(info_table.cell(i_ws, 0))
-            reduced_ws_name = 'VULCAN_%d_%d' % (self._reductionSetup.get_run_number(), ws_index)
-
-            # check whether the proposed-chopped workspace does exist
-            if AnalysisDataService.doesExist(reduced_ws_name):
-                chopped_ws_name_list.append(reduced_ws_name)
-            else:
-                # there won't be a workspace produced if there is no neutron event within the range.
-                if ws_index in target_chop_index_list:
-                    message += 'Reduced workspace %s does not exist. Investigate it!\n' % reduced_ws_name
-                    everything_is_right = False
-                else:
-                    message += 'Input reduced workspace %s does not exist, maybe no neutron.\n' % reduced_ws_name
-                continue
-
-            # convert unit and save for VULCAN-specific GSAS
-            tof_ws_name = "VULCAN_%d_TOF" % self._reductionSetup.get_run_number()
-            mantidsimple.ConvertUnits(InputWorkspace=reduced_ws_name,
-                                      OutputWorkspace=tof_ws_name,
-                                      Target="TOF",
-                                      EMode="Elastic",
-                                      AlignBins=False)
-
-            # overwrite the original file
-            vdrive_bin_ws_name = reduced_ws_name
-
-            # it might be tricky to give out the name of GSAS
-            if use_special_name:
-                gsas_file_name = os.path.join(chop_dir, '%d_%d.gda' % (self._reductionSetup.get_run_number(),
-                                                                       ws_index+1))
-            else:
-                gsas_file_name = os.path.join(chop_dir, '%d.gda' % (ws_index + 1))
-
-            # save to VULCAN GSAS and add a property as Note
-            mantidsimple.SaveVulcanGSS(InputWorkspace=tof_ws_name,
-                                       BinFilename=self._reductionSetup.get_vulcan_bin_file(),
-                                       OutputWorkspace=vdrive_bin_ws_name,
-                                       GSSFilename=gsas_file_name,
-                                       IPTS=self._reductionSetup.get_ipts_number(),
-                                       GSSParmFilename="Vulcan.prm")
-
-            # Add special property to output workspace
-            final_ws = AnalysisDataService.retrieve(vdrive_bin_ws_name)
-            final_ws.getRun().addProperty('VDriveBin', True, replace=True)
-
-            # update message
-            message += '%d-th: %s\n' % (i_ws, gsas_file_name)
-
-        # END-FOR
-
-        return everything_is_right, message, chopped_ws_name_list
-
-    def get_sub_splitters(self, split_start_index, split_stop_index):
-        """
-        chop splitters workspace to sub one
-        :param split_stop_index:
-        :return:
-        """
-        # get splitting workspace
-        split_ws_name, info_ws_name = self._reductionSetup.get_splitters(throw_not_set=True)
-        split_ws = AnalysisDataService.retrieve(split_ws_name)
-        sub_split_ws_name = split_ws.name() + '_{0}'.format(split_start_index)
-
-        # split
-        if isinstance(split_ws, mantid.dataobjects.SplittersWorkspace):
-            # splitters workspace
-            raise RuntimeError('It is not supported to split on a SplittersWorkspace')
-
-        elif isinstance(split_ws, mantid.api.MatrixWorkspace):
-            # Matrix workspace
-            # TODO/TEST - Need to test
-            vec_x = split_ws.readX(0)[split_start_index:split_stop_index+1]
-            vec_y = split_ws.readY(0)[split_start_index:split_stop_index]
-            vec_e = split_ws.readE(0)[split_start_index:split_stop_index]
-
-            mantidsimple.CreateWorkspace(DataX=vec_x, DataY=vec_y, DataE=vec_e, NSpec=1,
-                                         OutputWorkspace=sub_split_ws_name)
-
-        elif isinstance(split_ws, mantid.api.ITableWorkspace):
-            # Table workspace
-            # TODO/TEST - Need to verify
-            mantidsimple.CreateEmptyTableWorkspace(OutputWorkspace=sub_split_ws_name)
-            sub_split_ws = AnalysisDataService.retrieve(sub_split_ws_name)
-            sub_split_ws.addColumn('float', 'start')
-            sub_split_ws.addColumn('float', 'stop')
-            sub_split_ws.addColumn('str', 'index')
-
-            num_rows = split_ws.rowCount()
-            for i_row in range(split_start_index, min(split_stop_index, num_rows)):
-                start_time = split_ws.cell(i_row, 0)
-                stop_time = split_ws.cell(i_row, 1)
-                target = split_ws.cell(i_row, 2)
-                sub_split_ws.addRow([start_time, stop_time, target])
-
-        else:
-            # unsupported format
-            raise RuntimeError('Splitting workspace of type {0} is not supported.'.format(split_ws))
-
-        return sub_split_ws_name
-
     # def chop_and_reduce_large_output(self, split_ws_name, split_info_table):
     #     """
     #     Chop and reduce in the special case of large amount of output.
@@ -1663,29 +1501,6 @@ class ReduceVulcanData(object):
         else:
             shutil.copy(source_gsas_file_name, target_directory)
             os.chmod(target_file_name, 0664)
-
-        return
-
-    def execute_chop_reduction(self, clear_workspaces=True):
-        """
-        Execute the chopping and reduction including exporting the log files with chopped data
-        :return:
-        """
-        # create output directory
-        self.create_chop_dir()
-
-        # chop and reduce
-        status, message, output_ws_list = self.chop_reduce(self._choppedDataDirectory)
-
-        # create the log files
-        self.generate_sliced_logs(output_ws_list, self._chopExportedLogType)
-
-        # clear workspace? or later
-        if clear_workspaces:
-            for ws_name in output_ws_list:
-                mantidsimple.DeleteWorkspace(Workspace=ws_name)
-        else:
-            self._reducedWorkspaceList.extend(output_ws_list)
 
         return
 
@@ -2321,33 +2136,6 @@ class ReduceVulcanData(object):
                                                                           end_file_name)
 
         return
-
-    @staticmethod
-    def get_number_chopped_ws(split_ws_name):
-        """
-        get the number of expected chopped workspaces from splitters workspace
-        :param split_ws_name:
-        :return:
-        """
-        split_ws = AnalysisDataService.retrieve(split_ws_name)
-
-        if isinstance(split_ws, mantid.api.ITableWorkspace):
-            # table workspace
-            num_rows = split_ws.rowCount()
-            target_set = set()
-            for i_row in range(num_rows):
-                target = split_ws.cell(i_row, 2)
-                target_set.add(target)
-        else:
-            # matrix workspace case
-            target_set = set()
-            num_splitters = split_ws.readY(0)
-            for y in split_ws.readY(0):
-                int_y = int(y + 0.1)
-                target_set.add(int_y)
-        # END-FOR
-
-        return len(target_set)
 
     def get_reduced_files(self):
         """
